rename class from RequestHandlers to RequestErrors
